split package graph into multiple submodules
